Separate maker button by board. Includes tests
